% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featExt.R
\name{featExt}
\alias{featExt}
\title{Performing two-stage feature extraction}
\usage{
featExt(b, d, L, tau, alpha, stage1_thres, stage2_thres)
}
\arguments{
\item{b}{\eqn{L \times \tau} matrix of estimates of beta; posterior means of MCMC samples}

\item{d}{\eqn{L \times \tau} matrix of estimates of auxiliary variable d; posterior means of MCMC samples}

\item{L}{number of locations}

\item{tau}{number of time points}

\item{alpha}{significance level used to compute chi-squared threshold in second stage; controls selection rate}

\item{stage1_thres}{threshold in first stage to induce sparsity; fixed quantity chosen sufficiently small}

\item{stage2_thres}{threshold in second stage; chi-squared upper (1 - alpha)-quantile}
}
\value{
list containing nonzero estimates after first stage, active indices according to the second stage
}
\description{
Performing two-stage feature extraction
}
