% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genX.R
\name{genX}
\alias{genX}
\title{Function to Generate Data Array}
\usage{
genX(y, beta, n, N, L, tau, mu, rho, S)
}
\arguments{
\item{y}{binary response vector}

\item{beta}{coefficient matrix}

\item{n}{number of subjects}

\item{N}{value chosen sufficiently larger than n to sample X from}

\item{L}{number of locations}

\item{tau}{number of time points}

\item{mu}{mean of MVN distribution to generate U}

\item{rho}{like a correlation coefficient used to generate S}

\item{S}{standard deviation of MVN distribution to generate U}
}
\value{
three-dimensional array of data
}
\description{
Function to Generate Data Array
}
