% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Subject Level Prediction Using Length of Longest Run}
\usage{
pred(b_ests, bInd, X_test)
}
\arguments{
\item{b_ests}{\eqn{L \times \tau} matrix of coefficient estimates from two-stage feature extraction}

\item{bInd}{indices of selected locations from two-stage feature extraction}

\item{X_test}{\eqn{m \times L \times \tau} dimensional array of test data}
}
\value{
binary predicted response vector of length \eqn{m}
}
\description{
Subject Level Prediction Using Length of Longest Run
}
