% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predWeighted.R
\name{predWeighted}
\alias{predWeighted}
\title{Weighted Subject-Level Prediction}
\usage{
predWeighted(b_ests, bInd, X_test, thres)
}
\arguments{
\item{b_ests}{\eqn{L \times \tau} matrix of coefficient estimates from two-stage feature extraction}

\item{bInd}{indices of selected locations from two-stage feature extraction}

\item{X_test}{\eqn{L \times \tau} dimensional matrix of test data for subject i}

\item{thres}{threshold used to make final prediction for subject i}
}
\value{
list containing subject-level predictions, local probabilities, local weights, and weighted predictions
}
\description{
Component-wise products of the local weights and local prediction probabilities
}
