% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procSim.R
\name{procSim}
\alias{procSim}
\title{Process MCMC samples for an individual replication}
\usage{
procSim(
  beta,
  b_samps,
  y_test,
  X_test,
  actInd,
  stage1_thres = 10^(-3),
  FDRc = 0.05
)
}
\arguments{
\item{beta}{# true coefficient vector}

\item{b_samps}{# MCMC samples obtained from GD model}

\item{y_test}{# response vector - test set}

\item{X_test}{# data array - test set}

\item{actInd}{# indices of predetermined active locations}

\item{stage1_thres}{threshold in first stage to induce sparsity; fixed quantity chosen sufficiently small}

\item{FDRc}{FDR level at which the FDR is controlled - default \eqn{0.05}}
}
\value{
coefficient estimates, selected locations, estimation and prediction error rates, proportion of correct and incorrect indices, subject-level predictions, ROC, AUC, best threshold
}
\description{
Compute estimation rates, proportion of correct and incorrect indices, prediction error rates, and AUC
}
