% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_est.R
\name{rates_est}
\alias{rates_est}
\title{Function to Compute Error Rates of Estimation}
\usage{
rates_est(x, x_hat)
}
\arguments{
\item{x}{# vector containing true values of the coefficients}

\item{x_hat}{# vector containing predicted values of the coefficients}
}
\value{
vector containing the rmse, tpr, and fpr of estimation
}
\description{
Function to Compute Error Rates of Estimation
}
