% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_pred.R
\name{rates_pred}
\alias{rates_pred}
\title{Function to Compute Error Rates of Prediction}
\usage{
rates_pred(x, x_hat)
}
\arguments{
\item{x}{vector containing true values of the coefficients}

\item{x_hat}{vector containing predicted values of the coefficients}
}
\value{
vector containing the tpr, fpr, and pe of predictions
}
\description{
Function to Compute Error Rates of Prediction
}
