% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemSum.R
\name{riemSum}
\alias{riemSum}
\title{Riemman Sums of Location Vectors}
\usage{
riemSum(x)
}
\arguments{
\item{x}{\eqn{\tau \times 1} local vector of coefficient estimates from FDR in stage 1}
}
\value{
Riemman sum of the given location
}
\description{
Riemman Sums of Location Vectors
}
