% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEEG.R
\name{runEEG}
\alias{runEEG}
\title{Title}
\usage{
runEEG(
  y,
  X,
  tau0 = 10^(-5),
  n_trn = 100,
  alpha = 0.05,
  stage1_thres = 10^(-3),
  ncores = detectCores(),
  chains = 1,
  warmup = 300,
  iter = 1000,
  seed = 1234,
  stan_seed = 2411
)
}
\arguments{
\item{y}{binary response vector corresponding to \eqn{n} subjects}

\item{X}{\eqn{n \times L \times \tau} dimensional tensor of EEG data for \eqn{n} subjects, \eqn{L} locations, and \eqn{\tau} time points}

\item{tau0}{hyperparameter used in GD prior}

\item{n_trn}{size of training set}

\item{alpha}{significance level used to compute chi-squared threshold in second stage; controls selection rate}

\item{stage1_thres}{threshold in first stage to induce sparsity; fixed quantity chosen sufficiently small}

\item{ncores}{number of cores to use when fitting local models in parallel}

\item{chains}{number of chains to use in MCMC sampling}

\item{warmup}{number of warmup iterations to use in MCMC sampling}

\item{iter}{number of total iterations in MCMC sampling}

\item{seed}{random seed for reproducibility}

\item{stan_seed}{random seed used in stan code}
}
\value{
list containing important measures to report from the GD fitted model
}
\description{
Title
}
