% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_beta.R
\name{true_beta}
\alias{true_beta}
\title{Generate true beta coefficients}
\usage{
true_beta(
  L,
  tau,
  actL = ceiling(0.2 * L),
  actL_p_t = 0.9,
  inactL_p_t = 0.1,
  c = 3
)
}
\arguments{
\item{L}{number of locations}

\item{tau}{number of time points}

\item{actL}{number of active locations - default \eqn{actL = ceiling(0.2*L)}}

\item{actL_p_t}{prob of coefficient at time point t being significant for an active location - default \eqn{actL_p_{lt} = 0.9}}

\item{inactL_p_t}{prob of coefficient at time point t being significant for an inactive location - default \eqn{inactL_p_{lt} = 0.1}}

\item{c}{multiplying constant to increase the magnitude of nonzero entries in the true beta matrix - default \eqn{c = 3}}
}
\value{
list containing \eqn{L \times \tau} matrix of beta coefficients and vector of indices of active locations
}
\description{
Active locations should display a longer streaky behavior of 1's over time
}
