% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMods.R
\name{localMods}
\alias{localMods}
\title{Constructing Local Models in Parallel}
\usage{
localMods(y, X, tau0 = 10^(-5), chains = 4, warmup = 3000, iter = 10000)
}
\arguments{
\item{y}{binary response correspoding to \eqn{n} subjects}

\item{X}{\eqn{n} x \eqn{L} x \eqn{tau} - tensor (EEG data) for \eqn{n} subjects, \eqn{L} locations and \eqn{tau} time points}

\item{tau0}{predetermined constant \eqn{\tau0} - default \eqn{\tau0 = 10^(-5)}}

\item{chains}{number of chains in MCMC algorithm - default \eqn{chains = 4}}

\item{warmup}{number of iterations to throw away as burnin at the start of the mcmc run - default \eqn{warmup = 3000}}

\item{iter}{number of iterations to run in each mcmc chain - default \eqn{iter = 10000}}
}
\value{
list containing posterior mean estimates of the coefficients and MCMC samples of the coefficients
}
\description{
Constructing Local Models in Parallel
}
