% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSim.R
\name{runSim}
\alias{runSim}
\title{Function to Run GD Simulations}
\usage{
runSim(
  y,
  X,
  beta,
  trn_ind,
  tau0 = 10^(-5),
  chains = 4,
  warmup = 3000,
  iter = 10000,
  init_thres = 1e-04,
  bf_thres = 0.1
)
}
\arguments{
\item{y}{response vector}

\item{X}{data}

\item{beta}{true coefficient vector}

\item{trn_ind}{training indices}

\item{tau0}{hyperparameter used in GD prior}

\item{chains}{number of MCMC chains}

\item{warmup}{number of warmup iterations}

\item{iter}{total number of iterations}

\item{init_thres}{inital threshold in two stage feature extraction}

\item{bf_thres}{BF threshold in two stage feature extraction}
}
\value{
list containing important measures to report from simulation
}
\description{
Function to Run GD Simulations
}
