% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSim.R
\name{runSim}
\alias{runSim}
\title{Simulations - Fit Local Models in Parallel Using GD Prior}
\usage{
runSim(
  n = 130,
  L,
  tau,
  n_trn,
  N = 2 * n,
  tau_start,
  tau_end,
  modRstan,
  actL = ceiling(0.2 * L),
  actL_p_t = 0.8,
  inactL_p_t = 0.2,
  c = 5,
  rho = 0.5,
  tau0 = 10^(-5),
  chains = 1,
  warmup = 3000,
  iter = 10000,
  stan_seed = 2411
)
}
\arguments{
\item{n}{number of subjects}

\item{L}{number of locations}

\item{tau}{number of time points}

\item{n_trn}{size of training set}

\item{N}{value chosen sufficiently larger than n to sample X from}

\item{tau_start}{index in vector of time points representing the start of the current subgroup}

\item{tau_end}{index in vector of time points representing the end of the current subgroup}

\item{modRstan}{model code declaring the stan model}

\item{actL}{number of active locations - default \eqn{actL = ceiling(0.2*L)}}

\item{actL_p_t}{prob of coefficient at time point t being significant for an active location - default \eqn{actL_p_{lt} = 0.8}}

\item{inactL_p_t}{prob of coefficient at time point t being significant for an inactive location - default \eqn{inactL_p_{lt} = 0.2}}

\item{c}{multiplying constant to increase the magnitude of nonzero entries in the true beta matrix - default \eqn{c = 3}}

\item{rho}{used to generate covariance matrix when generating data - controls degree of multicollinearity}

\item{tau0}{hyperparameter used in GD prior}

\item{chains}{number of MCMC chains}

\item{warmup}{number of warmup iterations}

\item{iter}{total number of iterations}

\item{stan_seed}{seed for running stan model}
}
\value{
list containing important measures to report from simulation
}
\description{
Simulations - Fit Local Models in Parallel Using GD Prior
}
